[Refactor] update mapped API endpoints
